+++ Editor/TagTypeGenerator.cs
@@ -13,11 +13,11 @@ using static System.String;
 
 namespace AlkimeeGames.TagLayerTypeGenerator.Editor
 {
-    /// <summary>Generates a file containing a class; which contains constant string definitions for each tag in the project.</summary>
+    /// <summary>Generates a file containing a type; which contains constant string definitions for each tag in the project.</summary>
     public sealed class TagTypeGenerator : TypeGenerator<TagTypeGenerator>
     {
-        /// <summary>Used to check what tag strings are in the Tag class.</summary>
-        private readonly HashSet<string> _inClass = new HashSet<string>();
+        /// <summary>Used to check what tag strings are in the Tag type.</summary>
+        private readonly HashSet<string> _inType = new HashSet<string>();
 
         /// <summary>Used to check what tag strings are in the project.</summary>
         private readonly HashSet<string> _inUnity = new HashSet<string>();
@@ -25,7 +25,7 @@ namespace AlkimeeGames.TagLayerTypeGenerator.Editor
         /// <summary>The absolute path to the file containing the tags.</summary>
         [NotNull] private static string TagFilePath => $"{Application.dataPath}/{Settings.Tag.FilePath}";
 
-        /// <summary>Used to read the values from the Class. If we don't use reflection to find the Class, we tie ourselves to a specific configuration which isn't ideal.</summary>
+        /// <summary>Used to read the values from the type. If we don't use reflection to find the type, we tie ourselves to a specific configuration which isn't ideal.</summary>
         [CanBeNull] private static Type TagType => Type.GetType($"{Settings.Tag.Namespace}.{Settings.Tag.TypeName}, {Settings.Tag.Assembly}");
 
         /// <summary>Configures the callback for when the editor sends a message the project has changed.</summary>
@@ -58,8 +58,8 @@ namespace AlkimeeGames.TagLayerTypeGenerator.Editor
             return false;
         }
 
-        /// <summary>Checks if the values defined in the class are the same as in Unity itself.</summary>
-        /// <returns>True if the tags in the project don't match the tags in the class.</returns>
+        /// <summary>Checks if the values defined in the type are the same as in Unity itself.</summary>
+        /// <returns>True if the tags in the project don't match the tags in the type.</returns>
         private bool HasChangedTags()
         {
             _inUnity.Clear();
@@ -67,14 +67,14 @@ namespace AlkimeeGames.TagLayerTypeGenerator.Editor
             foreach (string tag in InternalEditorUtility.tags)
                 _inUnity.Add(tag.Replace(" ", Empty));
 
-            _inClass.Clear();
+            _inType.Clear();
 
             FieldInfo[] fields = TagType.GetFields(BindingFlags.Public | BindingFlags.Static);
             foreach (FieldInfo fieldInfo in fields)
                 if (fieldInfo.IsLiteral)
-                    _inClass.Add(fieldInfo.Name);
+                    _inType.Add(fieldInfo.Name);
 
-            return !_inClass.SetEquals(_inUnity);
+            return !_inType.SetEquals(_inUnity);
         }
 
         /// <summary>Validates if we can generate a new tags file.</summary>
@@ -88,7 +88,7 @@ namespace AlkimeeGames.TagLayerTypeGenerator.Editor
             return true;
         }
 
-        /// <summary>Generates a new Tags class file.</summary>
+        /// <summary>Generates a new Tags type file.</summary>
         public override void GenerateFile()
         {
             // Start with a compileUnit to create our code and give it an optional namespace.
@@ -100,34 +100,23 @@ namespace AlkimeeGames.TagLayerTypeGenerator.Editor
             ValidateIdentifier(codeNamespace, Settings.Tag.Namespace);
 
             // Declare a type that is public and sealed.
-            var typeDeclaration = new CodeTypeDeclaration(Settings.Tag.TypeName)
-            {
-                IsClass = true,
-                TypeAttributes = TypeAttributes.Public | TypeAttributes.Sealed
-            };
-
-            // Validate the type name.
-            ValidateIdentifier(typeDeclaration, Settings.Tag.TypeName);
-
-            // Add some comments so the class describes it's intended usage.
-            AddComments(typeDeclaration);
-
-            // Create members in the type for each tag in the project.
-            CreateTagMembers(typeDeclaration);
+            var tagType = new CodeTypeDeclaration(Settings.Tag.TypeName) {IsClass = true, TypeAttributes = TypeAttributes.Public | TypeAttributes.Sealed};
+            ValidateIdentifier(tagType, Settings.Tag.TypeName);
 
             // Add the type declaration to the namespace.
-            codeNamespace.Types.Add(typeDeclaration);
+            codeNamespace.Types.Add(tagType);
+
+            // Add some comments so the type to describes it's intended usage.
+            AddComments(tagType);
+
+            // Create members in the type for each tag in the project.
+            CreateTagMembers(tagType);
 
             // With a StringWriter and a CSharpCodeProvider; generate the code.
-
             using (var stringWriter = new StringWriter())
             {
                 using (var codeProvider = new CSharpCodeProvider())
-                    codeProvider.GenerateCodeFromCompileUnit(compileUnit, stringWriter, new CodeGeneratorOptions
-                    {
-                        BracingStyle = "C",
-                        BlankLinesBetweenMembers = false
-                    });
+                    codeProvider.GenerateCodeFromCompileUnit(compileUnit, stringWriter, new CodeGeneratorOptions {BracingStyle = "C", BlankLinesBetweenMembers = false});
 
                 // Create the asset path if it doesn't already exist.
                 CreateAssetPathIfNotExists(TagFilePath);
@@ -141,36 +130,30 @@ namespace AlkimeeGames.TagLayerTypeGenerator.Editor
             InvokeOnFileGeneration();
         }
 
-        /// <summary>Adds a verbose comment (like this one) to the class.</summary>
+        /// <summary>Adds a verbose comment (like this one) to the type.</summary>
         /// <param name="typeDeclaration">The <see cref="CodeTypeDeclaration" /> to add the comment to.</param>
         private void AddComments([NotNull] CodeTypeMember typeDeclaration)
         {
             var commentStatement = new CodeCommentStatement(
-                "<summary>\r\n Use these string constants when comparing tags in code / scripts.\r\n </summary>" +
-                "\r\n <example>\r\n <code>\r\n if (other.gameObject.CompareTag(Tags.Player)) {\r\n     Destroy(other.gameObject);" +
-                "\r\n }\r\n </code>\r\n </example>",
+                "<summary>\r\n Use these string constants when comparing tags in code / scripts.\r\n </summary>\r\n <example>\r\n <code>\r\n if " +
+                $"(other.gameObject.CompareTag({Settings.Tag.TypeName}.Player)) {{\r\n     Destroy(other.gameObject);\r\n }}\r\n </code>\r\n </example>",
                 true);
 
             typeDeclaration.Comments.Add(commentStatement);
         }
 
-        /// <summary>Creates members for each tag in the project and adds them to the <paramref name="typeDeclaration" />.</summary>
-        /// <param name="typeDeclaration">The <see cref="CodeTypeDeclaration" /> to add the tag members to.</param>
-        private void CreateTagMembers(CodeTypeDeclaration typeDeclaration)
+        /// <summary>Creates members for each tag in the project and adds them to the <paramref name="tagType" />.</summary>
+        /// <param name="tagType">The <see cref="CodeTypeDeclaration" /> to add the tag members to.</param>
+        private void CreateTagMembers(CodeTypeDeclaration tagType)
         {
             foreach (string tag in InternalEditorUtility.tags)
             {
-                var field = new CodeMemberField
-                {
-                    Attributes = MemberAttributes.Public | MemberAttributes.Const,
-                    Name = tag.Replace(" ", Empty),
-                    Type = new CodeTypeReference(typeof(string)),
-                    InitExpression = new CodePrimitiveExpression(tag)
-                };
+                var field = new CodeMemberField(typeof(string), tag.Replace(" ", Empty))
+                    {Attributes = MemberAttributes.Public | MemberAttributes.Const, InitExpression = new CodePrimitiveExpression(tag)};
 
                 ValidateIdentifier(field, tag);
 
-                typeDeclaration.Members.Add(field);
+                tagType.Members.Add(field);
             }
         }
     }