+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2021-04-03
+
+- Generate types for both tags and layers (layers were previous enums).
+- Introduced TagAttribute and LayerAttribute which can be applied to any 'int' or 'string' property.
+
 ## [1.0.7] - 2021-04-02
 
 - Improved validation.